fixed hours and updated views

+++ db/migrate/002_create_timecards.rb
@@ -3,7 +3,7 @@ class CreateTimecards < ActiveRecord::Migration[5.1]
     create_table :timecards do |t|
       t.belongs_to  :user
       t.datetime    :start_time
-      t.float       :hours
+      t.decimal     :hours
 
       t.timestamps
     end
